+++ src/helpers/fsHelper.cpp
@@ -1,8 +1,8 @@
-#include "fsHelper.h"
-#include "assert.hpp"
+#include "helpers/fsHelper.h"
+#include "helpers/assert.hpp"
+#include "helpers/memoryHelper.h"
 #include "init.h"
 #include "logger/Logger.hpp"
-#include "memoryHelper.h"
 #include "nx/abort.h"
 
 namespace FsHelper {
